refactor: 컴포넌트에서 style코드 분리 - S-dot 형식으로
