+++ src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import CardEnrollForm from "./components/CardEnrollForm";
+import CardEnrollForm from "./components/CardEnrollForm/CardEnrollForm";
 
 function App() {
   return (
