+++ src/components/CardEnrollForm/style.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const CardEnrollFormContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding-top: 77px;
+`;
+
+export const CardInformation = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  width: 376px;
+  padding: 45px 30px;
+  gap: 16px;
+`;
