+++ src/components/CardExpirationDate/style.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const CardDateContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  gap: 16px;
+`;
+
+export const CardDateBox = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+`;
+
+export const InputContainer = styled.div`
+  display: flex;
+  width: 100%;
+  gap: 10px;
+`;
