+++ src/components/CardNumbers/style.ts
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+export const CardNumbersContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+`;
+
+export const CardNumberBox = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+`;
+
+export const InputContainer = styled.div`
+  display: flex;
+  gap: 10px;
+`;
