+++ src/components/CardOwnerName/CardOwnerName.tsx
@@ -1,32 +1,13 @@
 import { useState } from "react";
-import Input from "./atoms/Input";
-import { TitleText, LabelText } from "./atoms/text";
-import styled from "styled-components";
-import ErrorMessage from "./ErrorMessage";
+import Input from "../atoms/Input";
+import { TitleText, LabelText } from "../atoms/text";
+import ErrorMessage from "../ErrorMessage/ErrorMessage";
 import {
   executeValidators,
   isInvalidNameLength,
   isInvalidOwnerName,
-} from "../utils/validators";
-
-const CardOwnerNameContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  gap: 16px;
-`;
-
-const CardOwnerNameBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-`;
-
-const InputContainer = styled.div`
-  display: flex;
-  width: 100%;
-  gap: 10px;
-`;
+} from "../../utils/validators";
+import * as S from "./style";
 
 interface Props {
   cardOwnerName: CardInfoValue;
@@ -65,11 +46,11 @@ export default function CardOwnerName({
   };
 
   return (
-    <CardOwnerNameContainer>
+    <S.CardOwnerNameContainer>
       <TitleText>카드 소유자 이름을 입력해 주세요</TitleText>
-      <CardOwnerNameBox>
+      <S.CardOwnerNameBox>
         <LabelText>소유자 이름</LabelText>
-        <InputContainer>
+        <S.InputContainer>
           <Input
             maxLength={15}
             placeholder="JOHN DOE"
@@ -78,9 +59,9 @@ export default function CardOwnerName({
             onChangeInput={(value) => onChangeNameInput(value)}
             onBlurInput={(value) => onBlurNameInput(value)}
           />
-        </InputContainer>
+        </S.InputContainer>
         <ErrorMessage message={errorMessage}></ErrorMessage>
-      </CardOwnerNameBox>
-    </CardOwnerNameContainer>
+      </S.CardOwnerNameBox>
+    </S.CardOwnerNameContainer>
   );
 }
