+++ src/components/ErrorMessage/ErrorMessage.tsx
@@ -1,0 +1,9 @@
+import * as S from "./style";
+
+interface Props {
+  message: string;
+}
+
+export default function ErrorMessage({ message }: Props) {
+  return <S.ErrorMessage>{message}</S.ErrorMessage>;
+}
